+++ crates/swc_css_codegen/src/lib.rs
@@ -2,6 +2,7 @@
 #![allow(clippy::needless_update)]
 
 pub use std::fmt::Result;
+use std::str::from_utf8;
 
 use swc_common::Spanned;
 use swc_css_ast::*;
@@ -862,7 +863,9 @@ where
     #[emitter]
     fn emit_str(&mut self, n: &Str) -> Result {
         if self.config.minify {
-            self.wr.write_str(Some(n.span), &n.value)?;
+            let minified = minify_string(&*n.value);
+
+            self.wr.write_raw(Some(n.span), &minified)?;
         } else {
             self.wr.write_raw(Some(n.span), &n.raw)?;
         }
@@ -2011,4 +2014,70 @@ fn minify_hex_color(value: &str) -> String {
     }
 
     value.to_ascii_lowercase()
+}
+
+fn minify_string(value: &str) -> String {
+    let mut minified = String::new();
+
+    let mut dq = 0;
+    let mut sq = 0;
+
+    for c in value.chars() {
+        match c {
+            // If the character is NULL (U+0000), then the REPLACEMENT CHARACTER (U+FFFD).
+            '\0' => {
+                minified.push('\u{FFFD}');
+            }
+            // If the character is in the range [\1-\1f] (U+0001 to U+001F) or is U+007F, the
+            // character escaped as code point.
+            '\x01'..='\x1F' | '\x7F' => {
+                static HEX_DIGITS: &[u8; 16] = b"0123456789abcdef";
+
+                let b3;
+                let b4;
+                let char_as_u8 = c as u8;
+
+                let bytes = if char_as_u8 > 0x0f {
+                    let high = (char_as_u8 >> 4) as usize;
+                    let low = (char_as_u8 & 0x0f) as usize;
+
+                    b4 = [b'\\', HEX_DIGITS[high], HEX_DIGITS[low], b' '];
+
+                    &b4[..]
+                } else {
+                    b3 = [b'\\', HEX_DIGITS[c as usize], b' '];
+
+                    &b3[..]
+                };
+
+                minified.push_str(from_utf8(bytes).unwrap());
+            }
+            // If the character is '"' (U+0022) or "\" (U+005C), the escaped character.
+            // We avoid escaping `"` to better string compression - we count the quantity of
+            // quotes to choose the best default quotes
+            '\\' => {
+                minified.push_str("\\\\");
+            }
+            '"' => {
+                dq += 1;
+
+                minified.push(c);
+            }
+            '\'' => {
+                sq += 1;
+
+                minified.push(c);
+            }
+            // Otherwise, the character itself.
+            _ => {
+                minified.push(c);
+            }
+        };
+    }
+
+    if dq > sq {
+        format!("'{}'", minified.replace('\'', "\\'"))
+    } else {
+        format!("\"{}\"", minified.replace('"', "\\\""))
+    }
 }
