+++ crates/swc_css_codegen/src/writer/mod.rs
@@ -14,8 +14,6 @@ pub trait CssWriter {
 
     fn write_raw(&mut self, span: Option<Span>, text: &str) -> Result;
 
-    fn write_str(&mut self, span: Option<Span>, text: &str) -> Result;
-
     fn write_raw_char(&mut self, span: Option<Span>, c: char) -> Result;
 
     fn write_newline(&mut self) -> Result;
